+++ tk/dbox/api.py
@@ -13,7 +13,7 @@ import typing as ty
 import base64
 
 from datetime import datetime as dt
-from .utils.types import WithMetaResponse
+from tk.dbox.utils.types import WithMetaResponse
 
 L = logging.getLogger(__name__)
 T = ty.TypeVar('T')
