+++ tk/dbox/arxiv.py
@@ -3,7 +3,8 @@ import dataclasses as dcls
 import typing as ty
 
 from html.parser import HTMLParser
-from .utils import types
+
+from tk.dbox.utils import types
 
 
 PDF_BASE = 'https://arxiv.com/pdf/{id}.pdf'
