+++ tk/dbox/utils/text.py
